Ensure tests are deterministic
It seems that tests were failing ~20% of the time, due to an
incomplete line stored in deadgrep--remaining-output.

Tests start multiple searches for "foo" without stopping the old
search. Ensure that when we initialise the search buffer, we terminate
any old process and reset the search state.

This seems to fixed the issue.