Add ability to specify a set of extra directories to search
Add the deadgrep-extra-searched-directories user-option.
- That's an association list:
  - key   := directory context : a string identifying a directory context
  - value := list of directories to search when the directory context is
             selected in the Directory field.

By default this is nil, and deadgrep continues to behave as before.

If deadgrep-extra-searched-directories holds something and inside there's a
key that is equal to the default-directory, then the ripgrep command line
is set to search the directory/ies identified in the list associated with the
alist elment value.

If no key match the default-directory, deadgrep uses "." as before and ends up
behaving as before.

This provides the ability to set up a list of directory contexts to pre-define
the directories that are effectively searched by deadgrep.  The real
advantage is that this list can contain a set of non-overlapping directories
to search (and the original directory if that's needed), something like /etc
and /usr for example.

It works with Tramp syntax as well, as long as the user reads the description
in the customization buffer when entering the paths of the extra directories.